% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{getDmatrix}
\alias{getDmatrix}
\title{getDmatrix}
\usage{
getDmatrix(data)
}
\arguments{
\item{data}{data table of  DNA methylation matrix.  Rownames are 'TargetID', Columns are 'Sample'}
}
\value{
D: D matrix, row is gene, column is sample
}
\description{
Getting D matrix for applying to methPLIER
}
