% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{getDMGs}
\alias{getDMGs}
\title{getDMGs}
\usage{
getDMGs(D, annot, threshold = 0.05, method = "fdr")
}
\arguments{
\item{D}{D matrix obtained by 'getDmatrix' function}

\item{annot}{tibble data frame as the sample annotation.}

\item{threshold}{q.value threshold (default is 0.05)}

\item{method}{method of p.adjust (default is 'fdr')}
}
\value{
result: tibble data frame
}
\description{
This function gets differentially methylated score genes
}
