% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{getDMGsInLV}
\alias{getDMGsInLV}
\title{getDMGsInLV}
\usage{
getDMGsInLV(
  methPLIER,
  D,
  LV,
  annot,
  frac = 0.8,
  threshold = 0.05,
  method = "fdr"
)
}
\arguments{
\item{methPLIER}{methPLIER object}

\item{D}{D matrix obtained by 'getDmatrix' function}

\item{LV}{integer of LV}

\item{annot}{sample annotation tibble dataframe. It have to contain following column,
AccessionNo: the name of samples,
cluster.sample: category of each samples.}

\item{frac}{A number specifying what percentage of the total weight of the genes to be extracted (default is 0.8)}

\item{threshold}{q.value threshold (default is 0.05)}

\item{method}{method of p.adjust (default is 'fdr')}
}
\value{
result: result of getDMGs function
}
\description{
Wrapper function of getDMGs for TopLVs
}
