% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{getDMPs}
\alias{getDMPs}
\title{getDMPs}
\usage{
getDMPs(data, annot, genes, threshold = 0.05, method = "fdr")
}
\arguments{
\item{data}{data frame of DNA methylation data.
The attribute "methylation" obtained by 'makeGvizObj' function can be used as arg.}

\item{annot}{tibble data frame as the sample annotation.
The attribute "cluster.sample" obtained by 'plotHeatmap' function can be used as arg.
  Column_1: 'cluster.sample'
  Column_2: 'AccessionNo'}

\item{genes}{vector of genes}

\item{threshold}{q.value threshold (default is 0.05)}

\item{method}{method of p.adjust (default is fdr)}
}
\value{
result: tibble data frame
}
\description{
This function gets differentially methylated probes
}
