% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{getGenesInLV}
\alias{getGenesInLV}
\title{getGenesInLV}
\usage{
getGenesInLV(methPLIER, LV, frac = 0.8)
}
\arguments{
\item{methPLIER}{methPLIER: methPLIER model}

\item{LV}{LV number of to be obtained pathway}

\item{frac}{A number specifying what percentage of the total weight of the genes to be extracted (default is 0.8)}
}
\value{
genes: vector of genes in LV
}
\description{
This function gets list of genes of LV
}
