% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{getNewDataB}
\alias{getNewDataB}
\title{getNewDataB}
\usage{
getNewDataB(D, plier.model)
}
\arguments{
\item{D}{a DNA methylation matrix, rows are genes, columns are samples}

\item{plier.model}{PLIER results, output from PLIER::PLIER()}
}
\value{
meth.new.b: a matrix that contains the values of each latent variable for each sample from the new dataset (D),
}
\description{
Apply a previously computed PLIER to a new dataset to get the LV x sample
matrix (B)
see main PLIER function:
https://github.com/wgmao/PLIER/blob/a2d4a2aa343f9ed4b9b945c04326bebd31533d4d/R/Allfuncs.R#L227
}
