% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{GetOrderedRowNorm}
\alias{GetOrderedRowNorm}
\title{GetOrderedRowNorm}
\usage{
GetOrderedRowNorm(D, plier.model)
}
\arguments{
\item{D}{a DNA methylation matrix, rows are genes, columns are samples}

\item{plier.model}{PLIER results, output from PLIER::PLIER()}
}
\value{
meth.cg: a row normalized (z-scored) DNA methylation matrix that matches the Z matrix of the PLIER model -- ordered to match, contains the same genes
}
\description{
Given a DNA methylation matrix of data that was not used to train a PLIER model
and the output of PLIER::PLIER, row-normalize, remove genes not in the
training data, set missing genes to zero (the mean), and reorder to match
plier.model$Z
This makes the input DNA methylation data suitable for projection into
the training latent variable space (see GetNewDataB) and for evaluating
reconstruction (MASE, correlation)
}
