% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{getPathway}
\alias{getPathway}
\title{getPathway}
\usage{
getPathway(genes, showCategory = 15)
}
\arguments{
\item{genes}{vector of genes in LV}

\item{showCategory}{number of plotting ontology in barplot}
}
\value{
list: edo, edox, edox2, barplot, dotplot, treeplot
}
\description{
This function gets pathway of top genes of LV
}
