% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{getTopLVs}
\alias{getTopLVs}
\title{getTopLVs}
\usage{
getTopLVs(cluster.LV, col.1, col.2, threshold = 0.05, method = "fdr")
}
\arguments{
\item{cluster.LV}{tibble talbe of LV cluster obtained by 'plotHeatmap' function}

\item{col.1}{column A}

\item{col.2}{column B}

\item{threshold}{threshold for q.value filtering (default is 0.05)}

\item{method}{method for p.adjust (default is 'fdr')}
}
\value{
res: tibble data frame
}
\description{
This function get top-LVs
}
