% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{makeGvizObj}
\alias{makeGvizObj}
\title{makeGvizObj}
\usage{
makeGvizObj(gene, data, genome = "hg38")
}
\arguments{
\item{gene}{vector of gene name, i.e. c('PGAP3', 'ERBB2')}

\item{data}{matrix of DNA methylation. Row as TargetID, Column as Sample. Only numeric (DNA methylation value)}

\item{genome}{the version of reference genome. (default: 'hg38')}
}
\value{
list of Gviz object
}
\description{
This function generate Gviz object
}
