% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{PLIERNewData}
\alias{PLIERNewData}
\title{methPLIER}
\usage{
PLIERNewData(D, C = "all", seed = 12345)
}
\arguments{
\item{D}{a DNA methylation matrix, rows are genes, columns are samples}

\item{C}{knowledge matrix}

\item{seed}{an integer to be supplied to set.seed() for reproducibility purposes, default is 12345}
}
\value{
plier.res: output from PLIER::PLIER()
}
\description{
methPLIER

A wrapper function for applying PLIER to a data set.
We use the following genesets that come with PLIER: bloodCellMarkersIRISDMAP, svmMarkers,
and canonicalPathways.
We set the k parameter for the PLIER model by
identifying the number of "significant PCs" with PLIER::num.pc and then
using sig PCs * 0.3. This is consistent with recommendations from the PLIER authors.
}
\references{
[Mao, W., et al., Nat. Meth. 16(7):607-610, (2019)]

[Taroni, N. J., et al., Cell Syst. 8(5):380-394.e4 (2019)]
}
\author{
Ken Takasawa
PLIERNewData
}
