% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{plotBoxplot.gviz}
\alias{plotBoxplot.gviz}
\title{plotBoxplot.gviz}
\usage{
plotBoxplot.gviz(
  gvizObj,
  sgroups,
  target.id,
  sizes = c(5, 1),
  fill = NULL,
  filename = NULL,
  dir = NULL,
  pfx = NULL,
  width = 10,
  height = 12,
  ...
)
}
\arguments{
\item{gvizObj}{Gviz object}

\item{sgroups}{vector of sample group}

\item{target.id}{vector of TargetID what you want to plot}

\item{sizes}{vector of sizes for detail Track and GeneRegion Track (default is c(5, 1))}

\item{fill}{vector of fill color for bwplot (default is NULL)}

\item{filename}{optional, filename for saving plot (default is NULL).
If you want to save plot, also 'dir' must be passed this function.}

\item{dir}{output directory path (default is NULL).
If you want to save plot, also 'filename' must be passed this function.}

\item{pfx}{optional, prefix for saving plot filenames (default is NULL)}

\item{width}{width of plot, numeric}

\item{height}{height of plot, numeric}

\item{...}{arguments for lattice::bwplot()}
}
\value{
plot box plot with Gviz
}
\description{
This function plots detail Track and GeneRegion Track
}
