% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{plotBoxplot}
\usage{
plotBoxplot(B, LV, cl)
}
\arguments{
\item{B}{B matrix obtained by 'getNewDataB' function}

\item{LV}{LV number of to be drawn}

\item{cl}{tibble data table. category of sample. "AccessionNo", and "cluster.sample" column needed.}
}
\value{
plot box plot
}
\description{
This function plots Boxplot
}
