% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{plotDetails}
\alias{plotDetails}
\title{plotDetails}
\usage{
plotDetails(identifier, methylation, sgroups, fill = NULL, ...)
}
\arguments{
\item{identifier}{TargetID of DNA methylation data.}

\item{methylation}{matrix of DNA methylation. Row as TargetID, Column as Sample. Only numeric (DNA methylation value)}

\item{sgroups}{vector of sample group}

\item{fill}{vector of fill color for bwplot (default is NULL)}

\item{...}{arguments for lattice::bwplot()}
}
\value{
plot bwplot
}
\description{
This function takes the output of gene name, methylation data, and sample group,
and plot methylation value on genomic position.
}
