% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{plotGvizObj}
\alias{plotGvizObj}
\title{plotGvizObj}
\usage{
plotGvizObj(
  gvizObj,
  type = c("a", "p"),
  filename = NULL,
  dir = NULL,
  pfx = NULL,
  ...
)
}
\arguments{
\item{gvizObj}{Gviz object}

\item{type}{vector for types of plotting by Gviz::plotTacks()}

\item{filename}{optional, filename for saving plot (default is NULL).
If you want to save plot, also 'dir' must be passed this function.}

\item{dir}{output directory path (default is NULL).
If you want to save plot, also 'filename' must be passed this function.}

\item{pfx}{optional, prefix for saving plot filenames (default is NULL)}

\item{...}{arguments for Gviz::plotTracks()}
}
\value{
plot Gviz track
}
\description{
This function plots Gviz object
}
