% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap}
\usage{
plotHeatmap(B, k = 2, k.lv = 4, annot.df = NULL, col = NULL)
}
\arguments{
\item{B}{B matrix obtained by 'getNewDataB' function}

\item{k}{number of cluster splitting samples (default is 2)}

\item{k.lv}{number of cluster splitting for LVs (default is 4)}

\item{annot.df}{optional. data frame of sample annotation. rowid is sample name.}

\item{col}{optional. list of annotaiton color. heatmap.col.RData as a test file.}
}
\value{
plot Heatmap and cluster table
}
\description{
This function plots Heatmap of B matrix with or without sample annotation
}
