% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{plotSurvival}
\alias{plotSurvival}
\title{plotSurvival}
\usage{
plotSurvival(
  cl,
  xlab = "Time Since Surgery (year)",
  ylab = "Relapce-Free Survival (probability)"
)
}
\arguments{
\item{cl}{data for survival analysis. Data must following columns.
"cluster" column: Cluster number to which the sample belongs. Each value should be numeric.
"Event" column: Event value for survival analysis. 0 or 1.
"Time" column: Time to event. Each value should be numeric.}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}
}
\value{
plot survival plot
}
\description{
This function plots Survival plot
}
