% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{rowNorm}
\alias{rowNorm}
\title{rowNorm}
\usage{
rowNorm(x)
}
\arguments{
\item{x}{a DNA methylation matrix, rows are genes, columns are samples}
}
\value{
x: Normalized data matrix
}
\description{
Row normalization to apply PLIERNewData function.
Subtracting mean value of the row from data, and dividing by standard deviation value of the row.
}
