% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methplier.R
\name{plotSurvival}
\alias{plotSurvival}
\title{plotSurvival}
\usage{
plotSurvival(
  cl,
  xlab = "Time Since Surgery (year)",
  ylab = "Relapce-Free Survival (probability)"
)
}
\arguments{
\item{cl}{data for survival analysis. Data must following columns.}

\item{xlab}{label of x-axis}

\item{ylab}{label of y-axis}

\item{cluster}{numeric. Cluster number to which the sample belongs.}

\item{Time}{numeric. Time to event.}

\item{Event}{numeric. Event value for survival analysis.}
}
\value{
plot survival plot
}
\description{
This function plots Survival plot
}
